Don't try to kill rethink if you didn't start it.

+++ rethinkdb-net-test/Integration/IntegrationTestSetup.cs
@@ -27,7 +27,8 @@ namespace RethinkDb.Test
         [TearDown]
         public void Teardown()
         {
-            rethinkProcess.Kill();
+            if (rethinkProcess != null)
+                rethinkProcess.Kill();
         }
 
         private string GetRethinkPath()
